requirements.txt
docker>=6.1.3
pyyaml>=6.0
requests>=2.31.0